Create `SwiftUISyncTestHost` and `SwiftUISyncTestHostTests` target to test AutoOpen & AsyncOpen property wrapper against the test server
