+++ Configuration/SwiftUISyncTestHost.xcconfig
@@ -1,0 +1,4 @@
+#include "TestHost.xcconfig"
+
+INFOPLIST_FILE = Realm/Tests/SwifttUISyncTestHost/Info.plist;
+SUPPORTED_PLATFORMS = macosx;
