+++ Configuration/SwiftUISyncTestHostTests.xcconfig
@@ -1,0 +1,5 @@
+#include "SwiftUISyncTestHost.xcconfig"
+
+TEST_TARGET_NAME = SwiftUISyncTestHost;
+SDKROOT = macosx;
+SUPPORTED_PLATFORMS = macosx;
