+++ Package.swift
@@ -260,6 +260,8 @@ let package = Package(
                 "fileformat-pre-null.realm",
                 "mixed_tests.py",
                 "set_tests.py",
+                "SwifttUISyncTestHost",
+                "SwifttUISyncTestHostUITests"
             ],
             cxxSettings: testCxxSettings
         ),
