+++ Realm/ObjectServerTests/SwiftObjectServerTests.swift
@@ -29,28 +29,6 @@ import RealmSyncTestSupport
 import RealmTestSupport
 #endif
 
-class SwiftHugeSyncObject: Object {
-    @objc dynamic var _id = ObjectId.generate()
-    @objc dynamic var data: Data?
-
-    override class func primaryKey() -> String? {
-        return "_id"
-    }
-
-    class func create() -> SwiftHugeSyncObject {
-        let fakeDataSize = 1000000
-        return SwiftHugeSyncObject(value: ["data": Data(repeating: 16, count: fakeDataSize)])
-    }
-}
-
-extension User {
-    func configuration(testName: String) -> Realm.Configuration {
-        var config = self.configuration(partitionValue: testName)
-        config.objectTypes = [SwiftPerson.self, SwiftHugeSyncObject.self, SwiftTypesSyncObject.self]
-        return config
-    }
-}
-
 @available(OSX 10.14, *)
 @objc(SwiftObjectServerTests)
 class SwiftObjectServerTests: SwiftSyncTestCase {
@@ -390,26 +368,6 @@ class SwiftObjectServerTests: SwiftSyncTestCase {
     }
 
     // MARK: - Progress notifiers
-
-    let bigObjectCount = 2
-
-    func populateRealm(user: User, partitionValue: String) {
-        do {
-            let user = try logInUser(for: basicCredentials())
-            let config = user.configuration(testName: partitionValue)
-            let realm = try openRealm(configuration: config)
-            try! realm.write {
-                for _ in 0..<bigObjectCount {
-                    realm.add(SwiftHugeSyncObject.create())
-                }
-            }
-            waitForUploads(for: realm)
-            checkCount(expected: bigObjectCount, realm, SwiftHugeSyncObject.self)
-        } catch {
-            XCTFail("Got an error: \(error) (process: \(isParent ? "parent" : "child"))")
-        }
-    }
-
     func testStreamingDownloadNotifier() {
         do {
             let user = try logInUser(for: basicCredentials())
@@ -645,7 +603,7 @@ class SwiftObjectServerTests: SwiftSyncTestCase {
         let appConfig = AppConfiguration(baseURL: "http://localhost:5678",
                                          transport: AsyncOpenConnectionTimeoutTransport(),
                                          localAppName: nil, localAppVersion: nil)
-        let app = App(id: appId, configuration: appConfig)
+        let app = App(id: appId)
 
         let syncTimeoutOptions = SyncTimeoutOptions()
         syncTimeoutOptions.connectTimeout = 2000
