+++ Realm/Tests/SwifttUISyncTestHost/SwifttUISyncTestHostApp.swift
@@ -1,0 +1,39 @@
+////////////////////////////////////////////////////////////////////////////
+//
+// Copyright 2021 Realm Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+////////////////////////////////////////////////////////////////////////////
+
+import SwiftUI
+import RealmSwift
+
+@main
+struct App: SwiftUI.App {
+    var body: some Scene {
+        let view: AnyView = {
+            switch ProcessInfo.processInfo.environment["test_type"] {
+            case "async_open":
+                return AnyView(AsyncOpenView())
+            case "auto_open":
+                return AnyView(EmptyView())
+            default:
+                return AnyView(EmptyView())
+            }
+        }()
+        return WindowGroup {
+            view
+        }
+    }
+}
